[refactor] aggrement 패키지를 user에서 분리
