+++ src/main/kotlin/com/meokq/api/agreement/annotations/ExplainSaveAgreement.kt
@@ -1,4 +1,4 @@
-package com.meokq.api.user.annotaions
+package com.meokq.api.agreement.annotations
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
@@ -6,8 +6,8 @@ import io.swagger.v3.oas.annotations.media.ExampleObject
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 
 @Operation(
-    summary = "(IRA002) 약관동의 내역 조회",
-    description = "약관동의 내역 조회",
+    summary = "(IRA001) 약관동의 내역 저장",
+    description = "약관동의 내역 저장",
     //tags = ["User"]
 )
 @ApiResponse(
@@ -17,20 +17,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
         mediaType = "application/json",
         examples = [ExampleObject(value = """
 {
-  "size": 10,
-  "data": [
-    {
-      "agreementType": "PERSONAL_INFO_COLLECTION",
-      "version": 0,
-      "acceptYn": "Y",
-      "acceptDate": "2023-12-30 11:46:28"
-    }
-  ],
-  "total": 1,
-  "page": 0,
   "status": "OK",
   "message": "Your request has been processed successfully."
 }
                 """)])]
 )
-annotation class ExplainSelectAgreement()
+annotation class ExplainSaveAgreement()
