+++ src/main/kotlin/com/meokq/api/agreement/repository/AgreementRepository.kt
@@ -1,7 +1,7 @@
-package com.meokq.api.user.repository
+package com.meokq.api.agreement.repository
 
 import com.meokq.api.core.repository.BaseRepository
-import com.meokq.api.user.model.Agreement
+import com.meokq.api.agreement.model.Agreement
 
 interface AgreementRepository : BaseRepository<Agreement, String> {
     //fun findAll(spec: Specification<Agreement>?, pageable: Pageable): Page<Agreement>
