+++ src/main/kotlin/com/meokq/api/agreement/request/AgreementSearchDto.kt
@@ -1,7 +1,7 @@
-package com.meokq.api.user.request
+package com.meokq.api.agreement.request
 
 import com.fasterxml.jackson.annotation.JsonIgnore
-import com.meokq.api.user.enums.AgreementType
+import com.meokq.api.agreement.enums.AgreementType
 
 data class AgreementSearchDto(
     @JsonIgnore
